+++ lib/shopify-cli/script/sdk/as/as-pect.config.js
@@ -1,0 +1,20 @@
+module.exports = {
+    include: ["**/*.spec.ts"],
+    add: ["**/*.include.ts"],
+    flags: {
+        "--runtime": ["stub"]
+    },
+    disclude: [/node_modules/],
+    imports: {},
+    performance: {
+        enabled: false,
+        maxSamples: 10000,
+        maxTestRunTime: 5000,
+        reportMedian: true,
+        reportAverage: true,
+        reportStandardDeviation: false,
+        reportMax: true,
+        reportMin: true,
+    },
+    outputBinary: false,
+};
