+++ lib/shopify-cli/script/sdk/as/index.ts
@@ -1,0 +1,1 @@
+export * from "./runtime_types"
