+++ lib/shopify-cli/script/templates/ts/as-pect.config.js
@@ -2,7 +2,8 @@ module.exports = {
   include: ["**/*.spec.ts"],
   add: ["**/*.include.ts"],
   flags: {
-    "--runtime": ["stub"]
+    "--runtime": ["stub"],
+    "--lib": ["node_modules"]
   },
   disclude: [/node_modules/],
   imports: {},
